+++ bin/generate-sources.sh
@@ -15,7 +15,9 @@ rm -rf "${SCRIPT_PATH}"/cloud.yml || true
 rm -rf "${SCRIPT_PATH}"/influxdb-clients-apigen || true
 wget https://raw.githubusercontent.com/influxdata/openapi/master/contracts/oss.yml -O "${SCRIPT_PATH}/oss.yml"
 wget https://raw.githubusercontent.com/influxdata/openapi/master/contracts/cloud.yml -O "${SCRIPT_PATH}/cloud.yml"
+wget https://raw.githubusercontent.com/influxdata/openapi/master/contracts/invocable-scripts.yml -O "${SCRIPT_PATH}/invocable-scripts.yml"
 git clone --single-branch --branch master https://github.com/bonitoo-io/influxdb-clients-apigen "${SCRIPT_PATH}/influxdb-clients-apigen"
+mvn -f "$SCRIPT_PATH"/influxdb-clients-apigen/openapi-generator/pom.xml compile exec:java -Dexec.mainClass="com.influxdb.MergeContracts" -Dexec.args="$SCRIPT_PATH/oss.yml $SCRIPT_PATH/invocable-scripts.yml"
 mvn -f "$SCRIPT_PATH"/influxdb-clients-apigen/openapi-generator/pom.xml compile exec:java -Dexec.mainClass="com.influxdb.AppendCloudDefinitions" -Dexec.args="$SCRIPT_PATH/oss.yml $SCRIPT_PATH/cloud.yml"
 
 # Generate client
@@ -85,6 +87,12 @@ cp -r "${SCRIPT_PATH}"/generated/lib/influx_db2/models/label_mapping.rb "${SCRIP
 cp -r "${SCRIPT_PATH}"/generated/lib/influx_db2/models/label_response.rb "${SCRIPT_PATH}"/../apis/lib/influxdb2/apis/generated/models/
 cp -r "${SCRIPT_PATH}"/generated/lib/influx_db2/models/label_update.rb "${SCRIPT_PATH}"/../apis/lib/influxdb2/apis/generated/models/
 cp -r "${SCRIPT_PATH}"/generated/lib/influx_db2/models/labels_response.rb "${SCRIPT_PATH}"/../apis/lib/influxdb2/apis/generated/models/
+cp -r "${SCRIPT_PATH}"/generated/lib/influx_db2/models/scripts.rb "${SCRIPT_PATH}"/../lib/influxdb2/client/models/
+cp -r "${SCRIPT_PATH}"/generated/lib/influx_db2/models/script.rb "${SCRIPT_PATH}"/../lib/influxdb2/client/models/
+cp -r "${SCRIPT_PATH}"/generated/lib/influx_db2/models/script_create_request.rb "${SCRIPT_PATH}"/../lib/influxdb2/client/models/
+cp -r "${SCRIPT_PATH}"/generated/lib/influx_db2/models/script_invocation_params.rb "${SCRIPT_PATH}"/../lib/influxdb2/client/models/
+cp -r "${SCRIPT_PATH}"/generated/lib/influx_db2/models/script_language.rb "${SCRIPT_PATH}"/../lib/influxdb2/client/models/
+cp -r "${SCRIPT_PATH}"/generated/lib/influx_db2/models/script_update_request.rb "${SCRIPT_PATH}"/../lib/influxdb2/client/models/
 
 # copy supporting files
 cp -r "${SCRIPT_PATH}"/generated/lib/influx_db2/api_client.rb "${SCRIPT_PATH}"/../apis/lib/influxdb2/apis/generated/
@@ -96,6 +104,13 @@ sed -i 's/::API//' "${SCRIPT_PATH}"/../lib/influxdb2/client/models/health_check.
 sed -i 's/::API//' "${SCRIPT_PATH}"/../lib/influxdb2/client/models/dialect.rb
 sed -i 's/::API//' "${SCRIPT_PATH}"/../lib/influxdb2/client/models/query.rb
 sed -i 's/::API//' "${SCRIPT_PATH}"/../lib/influxdb2/client/models/delete_predicate_request.rb
+sed -i 's/::API//' "${SCRIPT_PATH}"/../lib/influxdb2/client/models/scripts.rb
+sed -i 's/::API//' "${SCRIPT_PATH}"/../lib/influxdb2/client/models/script.rb
+sed -i 's/::API//' "${SCRIPT_PATH}"/../lib/influxdb2/client/models/script_create_request.rb
+sed -i 's/::API//' "${SCRIPT_PATH}"/../lib/influxdb2/client/models/script_invocation_params.rb
+sed -i 's/::API//' "${SCRIPT_PATH}"/../lib/influxdb2/client/models/script_language.rb
+sed -i 's/::API//' "${SCRIPT_PATH}"/../lib/influxdb2/client/models/script_update_request.rb
+sed -i "s/'Object'/'Hash<String, Object>'/" "${SCRIPT_PATH}"/../lib/influxdb2/client/models/script_invocation_params.rb
 
 rm -rf "${SCRIPT_PATH}"/generated
 rm -rf "${SCRIPT_PATH}/influxdb-clients-apigen"