+++ lib/influxdb2/client/query_api.rb
@@ -23,7 +23,7 @@ require_relative 'flux_csv_parser'
 require 'json'
 
 module InfluxDB2
-  # The client of the InfluxDB 2.0 that implement Query HTTP API endpoint.
+  # The client of the InfluxDB 2.x that implement Query HTTP API endpoint.
   #
   class QueryApi < DefaultApi
     DEFAULT_DIALECT = InfluxDB2::Dialect.new(header: true, delimiter: ',', comment_prefix: '#',
@@ -76,7 +76,7 @@ module InfluxDB2
       uri = _parse_uri('/api/v2/query')
       uri.query = URI.encode_www_form(org: org_param)
 
-      _post_json(payload.to_body.to_json, uri)
+      _request_json(payload.to_body.to_json, uri)
     end
 
     def _generate_payload(query: nil, dialect: nil, params: nil)